[find] add `--limit` option to limit result size
Sometimes when querying a large amount of assets it's only necessary to have
a few (i.e. selecting a few hosts to test something on), but if querying
against a large pool that query can be slow. Adding a limit option allows for
that set to be trimmed down without having to wait for all results to be
returned