Add Ability to Query Remote Collins
This adds in the remoteLookup flag as seen in the asset API on
https://tumblr.github.io/collins/api.html. both r and R were taken
so I was just out of luck on that and Z came to mind but open to
other ideas. Possibly this is always the default behavior as well
and a config option should be added so you don't have to remember to
query all the collins.